refactor(mapping): simplify unmarshaller
Assign pointer to Param:
It turns, Param being pointer is useful because
it can be directly converted some interface
when a receiver for that method is pointer.

Removed type filling:
It turn out that filling type is not needed.
It was needed only for ObjectParam type.
But now that type is refactored to let type key to be blank if it is.