refactor(es_type, generate): use []byte, remove binary type
Refer to the document of std library. see: https://pkg.go.dev/encoding/json@go1.19.3#Marshal
The author(me) did not notice that it mentions that
it marshal []byte into base64 encoded string,
and it does not mention that in doc comment of Unmarshal.

If you needs base64 string in your code,
we can justify that there still be the Binary type.
But for simplicity I'm removing this.